1. Fix compilations issues from the previous commit. 2. Modify test to be "FLOAT" and not "DOUBLE" 3. Add new interface file which helps to add #include<> vs #include "".

+++ tests/library/poplar/poplar_matmul.py
@@ -41,7 +41,7 @@ def make_sdfg(dtype):
 
 @pytest.mark.poplar
 def test_poplar():
-    sdfg = make_sdfg(np.float64)
+    sdfg = make_sdfg(np.float32)
     sdfg.compile()
     print("Success!")
